+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- (Add) New option in admin to change for how long the cake icon will be on new members
+- (Add) New option in admin to change the label of new members to `New Member`
+- (Add) New translations to reflect the overall changes
+- (Change) Switched to dayjs
+
 ## [1.0.1] - 2021-03-25
 
 - (fix) translations were not working